docker-compose added

+++ API_REST/servicemanagement/src/main/java/com/edv/servicemanagement/components/user/domain/services/impl/UserServiceImpl.java
@@ -99,7 +99,9 @@ public class UserServiceImpl implements UserService {
         }
 
         if(file!=null){
-            fileService.delete(currentUser.getFile().getId());
+            if(currentUser.getFile()!=null) {
+                fileService.delete(currentUser.getFile().getId());
+            }
             fileService.create(user,file);
         }
